fix: variable assignment (#236)
To resolve the PSScriptAnalyzer warning "The variable 'version' is assigned but never used," the `$version` variable should be inside the string, not just assign it and then use its value in a string literal.

Variables inside single quotes are not expanded, so you need to use double quotes and embed the variable.

Signed-off-by: Ryan Johnson <ryan.johnson@broadcom.com>